+++ Runtime/Extensions/XMLExtension/Helper classes/Interface/ITextValue.cs
@@ -1,0 +1,5 @@
+namespace System.Xml {
+    public interface ITextValue {
+        XMLIRWText Text { get; set; }
+    }
+}