+++ Runtime/Extensions/XMLExtension/Helper classes/XMLIRW.cs
@@ -16,6 +16,7 @@ namespace System.Xml {
         protected XMLIRW(XMLIRW parent, string name) : this(parent, name, XmlNodeType.None) {}
         protected XMLIRW(string name, XmlNodeType type) : this(default, name, type) {}
         protected XMLIRW(string name) : this(default, name, XmlNodeType.None) {}
+        protected XMLIRW() {}
 
         public abstract void Dispose();
 #pragma warning restore CS1591 // O comentário XML ausente não foi encontrado para o tipo ou membro visível publicamente