+++ Runtime/Extensions/XMLExtension/Helper classes/XMLIRWDocType.cs
@@ -6,23 +6,25 @@ namespace System.Xml {
 #pragma warning disable CS1591 // O comentário XML ausente não foi encontrado para o tipo ou membro visível publicamente
         private bool disposedValue;
 
-        public override string Name { get; set; } = string.Empty;
-        public override XMLIRW Parent { get; set; } = default;
-        public XMLIRWValue PudID { get; private set;}
-        public XMLIRWValue SysID { get; private set;}
-        public XMLIRWValue SubSet { get; private set;}
+        public XMLIRWText PudID { get; private set; }
+        public XMLIRWText SysID { get; private set; }
+        public XMLIRWText SubSet { get; private set; }
         public override XmlNodeType Type { get; set; }
+        public override XMLIRW Parent { get; set; } = default;
+        public override string Name { get; set; } = string.Empty;
 
-        public XMLIRWDocType(XMLIRWElement parent, string name, XMLIRWValue pudid, XMLIRWValue sysid, XMLIRWValue subset) : base(parent, name, XmlNodeType.DocumentType) {
-            this.PudID = pudid;
-            this.SysID = sysid;
-            this.SubSet = subset;
+        [Obsolete("Use the XMLIRWDocType(XMLIRW, string, object, object, object) constructor.")]
+        public XMLIRWDocType(XMLIRW parent, string name, XMLIRWValue pudid, XMLIRWValue sysid, XMLIRWValue subset) {}
+        [Obsolete("Use the XMLIRWDocType(string, object, object, object) constructor.")]
+        public XMLIRWDocType(string name, XMLIRWValue pudid, XMLIRWValue sysid, XMLIRWValue subset) {}
+        
+        public XMLIRWDocType(XMLIRW parent, string name, object pudid, object sysid, object subset) : base(parent, name, XmlNodeType.DocumentType) {
+            PudID = new XMLIRWText(pudid);
+            SysID = new XMLIRWText(sysid);
+            SubSet = new XMLIRWText(subset);
         }
-        public XMLIRWDocType(string name, XMLIRWValue pudid, XMLIRWValue sysid, XMLIRWValue subset) : this(default, name, pudid, sysid, subset) {}
-        public XMLIRWDocType(XMLIRWElement parent, string name, object pudid, object sysid, object subset) :
-            this(parent, name, new XMLIRWValue(pudid), new XMLIRWValue(sysid), new XMLIRWValue(subset)) {}
         public XMLIRWDocType(string name, object pudid, object sysid, object subset) :
-            this(name, new XMLIRWValue(pudid), new XMLIRWValue(sysid), new XMLIRWValue(subset)) {}
+            this((XMLIRW)null, name, pudid, sysid, subset) {}
 
         ~XMLIRWDocType() => Dispose(disposing: false);